[models,test] fix: give a late-declared variant the trainable dtype
_mix_precision runs while the adapter is built and only sees the base components.
The extra variants are materialized afterwards, from PEFT's fp32 defaults, and
nothing cast them: a DMD2 checkpoint shipped a bf16 generator beside an fp32 fake
score. That is twice the memory, a different numerical path for the two halves of
the objective, and an RMSNorm whose weight no longer matches its activations, so
every norm the fake score feeds falls off the fused kernel and reports it once per
rank as a dtype mismatch warning.

+++ tests/models/test_component_variants.py
@@ -439,7 +439,9 @@ class _TinyRoleModule(torch.nn.Module):
 
 class _TinyRoleAdapter(BaseAdapter):
     def __init__(self, finetune_type: str) -> None:
-        self.model_args = SimpleNamespace(finetune_type=finetune_type)
+        self.model_args = SimpleNamespace(
+            finetune_type=finetune_type, trainable_parameters_dtype=torch.float32
+        )
         self.target_module_map = {"transformer": ["target"]}
         component = _TinyRoleModule()
         component.requires_grad_(False)
